minor adjustments to integers

+++ src/main/java/com/example/demo/DebugProg.java
@@ -3,8 +3,8 @@ package com.example.demo;
 public class DebugProg {
     public static void main(String args []) {
 
-        int x = (int) 3.145;
-        int y = (int) 3.64;
+        int x = 3.145;
+        int y = 3.64;
 
         System.out.println("pi is approximately " + x);
         System.out.println("My GPA was " + y);
